requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
bcrypt==4.0.1
python-dotenv==1.0.0
flask-cors==4.0.0
pytest==7.4.0
email-validator==2.0.0
requests==2.31.0